DPL: use timeframes in fly to rate limit analysis (#7185)

+++ Framework/Core/src/CommonServices.cxx
@@ -525,7 +525,7 @@ auto sendRelayerMetrics(ServiceRegistry& registry, DataProcessingStats& stats) -
   }
 
   if (stats.consumedTimeframes) {
-    monitoring.send(Metric{stats.consumedTimeframes, "consumed-timeframes"}.addTag(Key::Subsystem, Value::DPL));
+    monitoring.send(Metric{(uint64_t)stats.consumedTimeframes, "consumed-timeframes"}.addTag(Key::Subsystem, Value::DPL));
   }
 
   stats.lastSlowMetricSentTimestamp.store(stats.beginIterationTimestamp.load());
